Json utility for json specific operations
The commit introduces a new file to encapsulate JSON specific
functionalities under a new namespace.
These functionalities are moved from current utils.cpp file.

The change has been done to obtain better maintainability by
clubbing a kind of functionality into a separate namespace.
